Refactor code to use SessionManager class only
- A vending machine normally has a single user only operating it at a given point in time & hence there wouldn't be  a need for the Inheritance & a separate sub class
